refactor(webapp): migrate button/CopyButton to v2 CopyButton
Extend v2 CopyButton with iconType?: 'clipboard' | 'link' (Link2 icon
from lucide-react) so the link-copy use case in Logs is preserved.
Update all callsites (Logs/Operation/Show, SecretTextArea, TagsInput)
to import from v2 and delete the v1 source file.
