+++ packages/webapp/src/components-v2/ui/CopyButton.tsx
@@ -1,10 +1,10 @@
-import { Check, Copy } from 'lucide-react';
+import { Check, Copy, Link2 } from 'lucide-react';
 import { useEffect, useState } from 'react';
 
 import { Button } from './Button';
 import { cn } from '@/utils/utils';
 
-export const CopyButton: React.FC<{ text: string; disabled?: boolean }> = ({ text, disabled }) => {
+export const CopyButton: React.FC<{ text: string; disabled?: boolean; iconType?: 'clipboard' | 'link' }> = ({ text, disabled, iconType = 'clipboard' }) => {
     const [copied, setCopied] = useState(false);
     const [hasInteracted, setHasInteracted] = useState(false);
 
@@ -27,11 +27,12 @@ export const CopyButton: React.FC<{ text: string; disabled?: boolean }> = ({ tex
 
     // This avoids the animation from playing when the button is initially rendered
     const animationClass = hasInteracted ? 'animate-in zoom-in-45' : '';
+    const Icon = iconType === 'link' ? Link2 : Copy;
 
     return (
         <Button type="button" disabled={disabled} data-copied={copied} variant="ghost" size="icon" onClick={copyToClipboard} className="group">
             <Check className={cn('size-3.5 hidden group-data-[copied=true]:inline', animationClass)} />
-            <Copy className={cn('size-3.5 inline group-data-[copied=true]:hidden', animationClass)} />
+            <Icon className={cn('size-3.5 inline group-data-[copied=true]:hidden', animationClass)} />
         </Button>
     );
 };
